:art: Split `WokeConfig` impl into multiple files
